+++ src/control_plane/migration/postgres/mod.rs
@@ -1,0 +1,8 @@
+#[cfg(test)]
+mod tests;
+
+mod backup_postgres_database;
+mod postgres_backup_options;
+
+pub(crate) use backup_postgres_database::backup_postgres_database;
+pub(crate) use postgres_backup_options::PostgresBackupOptions;